Set source area for installer errors
Summary:
To exclude these from UBE-0.
Original plan was to set source_area based on presence of INSTALL tag and exclude that, however source area is only based on the best tag which is set to TIER0 if the installer categorizes something as Tier0 which takes precedence over INSTALL. Fixing this requires making sure best tag always starts with INSTALLER if it originates from the installer.

Reviewed By: IanChilds

Differential Revision: D69088225

fbshipit-source-id: 7b53f9359c5dc083a65fcae9532fafcb047ec5e2