Exit early from daemon if daemon data fails to initialize
Summary:
Currently if `DaemonState::init_data` returns an error the error is stored in `DaemonState.data` and the daemon continues running and allows clients to connect. Running commands with this daemon cause a stored error to be shown repeatedly until the user runs `buck2 kill`.

Apparently this was stored as a Result so that the client can connect and show a readable error, which should no longer be necessary as of the previous diff.

For now, this preserves the existing behavior of showing the following message if the daemon start up fails with an error not during `DaemonState::init_data`.
```
    Try running `buck2 kill` and your command afterwards.
    Alternatively, try running `rm -rf ~/.buck/buckd` and your command afterwards",
```
Not showing that message during this type of failure may or may not make sense but that message could probably be refined later once we have more data on what specific errors are returned from the daemon.

Reviewed By: iguridi

Differential Revision: D68655291

fbshipit-source-id: 85d027689811be5efc03f3a6a119df3a9bb4155d