Specify the default device state to use for buck tests
Summary:
If we don't pass the setup_preference_key to remote execution, the device we lease will be in a random state, which can affect flakiness. Fix it.

As D68185161 is pushed to all tiers, we now can update the bzl to pass the setup_preference_key to buck. Default to use A14 userdebug lkg build. But also provide a config to let user overrides.

Note: the side effect is the lease time will be longer because lab need to setup the devices. But this is the right thing to do.

Reviewed By: ywang226

Differential Revision: D68281753

fbshipit-source-id: 0e027cd9a257ef151f582c037460eae477480496