Log command report errors
Summary:
To improve measuring missing command failures (non-user failed builds with an entry in buck_results but not buck2_builds).

I think we should exclude `has_command_report==0` from the number of missing failures since this should only represent non-streaming commands, commands where command line args were rejected by clap, or --help commands, none of which should write invocation records. It's possible writing the command report should be present but fails to write but from the buck side it's much simpler to guarantee that it is written for commands that should write it.

The errors from the command report should be useful to diagnose missing commands. This may have some impact on retention for buck_results but it should be very small with these limited to non-user errors and truncated.

Currently this only looks at `isolation_dir == v2` since those are the only commands where we are currently tracking Tier 0 errors. We can change this later if needed but it requires parsing the command line args to find the isolation dir.

Reviewed By: Will-MingLun-Li

Differential Revision: D69029673

fbshipit-source-id: 8acefaf0a1d0b85c83b333e49e43779c15fdb0b8