Make artifact_dump.rs types owning
Summary:
In the original diff that added --output-hashes-file JakobDegen asked if the author can move that information into the build report. I want to make use of the hashes in Citadel, so I'd like to move it to the build report. The only place that uses this logic is the dumper itself, so making all that owning until I remove it in D69025542. I tried to make the build report store references, but that was a pain in the butt.

https://www.internalfb.com/diff/D49055340?dst_version_fbid=249975850999233&transaction_fbid=3578314329114700

Reviewed By: JakobDegen

Differential Revision: D69025761

fbshipit-source-id: 784fd5230d7ff6eeaf11dbb4cc3958c4c761a2a9