Migrate gen_docs.py from using buck2.sh to buck2.py
Summary:
This was the final callsite of `buck2.sh`, after a bit of debugging I found out the new `buck2.py` was handling the `--help` command and not propagating deeper.

The fix is just to not auto-generate the `--help` argument in the `buck2.py`, this will pass it through the buck2 inner command.

I could add a new flag for help on this command but it can be done on a separate diff.

Reviewed By: JakobDegen

Differential Revision: D66787380

fbshipit-source-id: c2533a133805f25cf298c3adf8710a5085e70afc