Make invocation record event creation infallible
Summary:
Noticed these ? will cause `send_it` to return None if `peak_..` fields are empty and we won't upload an invocation record.

Suspicious this may be the reason for missing invocation records for commands that fail in daemon start up, because these only get populated by a snapshot event which may not have happened yet.

Reviewed By: JakobDegen

Differential Revision: D69214620

fbshipit-source-id: 508f41a7adf9ad17064bcf630d13636f15ea2ac9