fix test_xxx timeout
Summary:
Initially I added `followlinks=True` as on MacOS we use eden symlink redirection. But this doesn't work on linux, as in .eden folder there is a symlink pointing to itself and it leads to infinite recursion.
We don't really need to traverse the whole project dir, it's enough to traverse `buck-out` only

Reviewed By: ezgicicek

Differential Revision: D69306814

fbshipit-source-id: d7afd6cec1c0962ee8ffb70c4fe91389f93ab9eb