completions: Explicitly set `clap::ValueHint::Other` on target patterns
Summary:
The way that default value hints are handled in zsh changed in https://github.com/clap-rs/clap/commit/e40168c2fd6e0d711d52699895ba475d178d5024 , which has caused breakage.

This is a bit ad-hoc, but I think that's ok, because which commands we support completions on is in general a bit ad-hoc. Making that more principled requires improvements to clap

Reviewed By: VladimirMakaev

Differential Revision: D69350230

fbshipit-source-id: 0ec9659783989df981cfd04963bb504e68070ff2