reduce graph traversal memory, use set not dict
Summary:
A dict has to store an extra 8 bytes `Value` pointer for every entry, whereas a set stores an instance of the zero-length `()` rust type.
Note: this depends on `set` returning keys in the order they are added. The current implementation of `set` in `starlark-rust` is a `SmallMap`, which is the same type a `dict` is built on.

Reviewed By: JakobDegen

Differential Revision: D68671724

fbshipit-source-id: 886d04557591753cc1cdbac13eecdb9f0b180644