Stop using a filelist for linker invocations
Summary:
We would like to switch to using `--start-lib --end-lib` style archives with our linker invocations, but currently use a filelist to pass some inputs to the linker. You cannot include `--start-lib` and `--end-lib` flags in a filelist, it must be literally just a list of files.

Prior to this change we categorized linker inputs somewhat arbitrarily into "filelist inputs" and everything else. The filelist was listed last on the command line. Now that we no longer have a filelist, we wish to maintain similar, or at least equally "effective" order but need a logical way to describe the categories.

Previously the order was:

(1) Everything not in the second category
(2) Regular object files and static archives

The order within each category is not explicitly defined and depends on the structure of the graph. This could be an issue if a regular object file appears after a regular archive in the filelist.

In deciding the new order, we consider what it is we really care about here. We want to minimize loading, and we want to minimize duplicating symbols between binaries.

To minimize loading, we should list anything lazy last after shared libraries and eager object files. This means the linker will prefer using a symbol from a shared library, rather than loading it from a lazy object file. It also means a lazy object file won't be loaded to provide a symbol an eager object file also provides.

Now we consider the question of minimizing symbol duplication. If a loaded object file and a shared library both provide symbol foo, we might save on overall binary size by binding to the foo in the shared library (in case another executable also uses foo). However, it turns out this isn't how it works. The linker will always prefer the definition from a **loaded (eager)** object file, rather than using a definition from a shared library. This means the order between shared libraries and **loaded** object files doesn't matter. All that matters is lazy object files come last after everything else.

This is exactly the order we define:

(1) Everything not in the second category
(2) Lazy object files (regular archives)

This is a slight change from before, effectively removing regular object files from the filelist. Here we see a diff of the effective order for an IG dylib: P1726859841. You see the regular object files moved upwards out of the filelist.

Reviewed By: rmaz

Differential Revision: D68593087

fbshipit-source-id: 432c7503180389380972482a30298d6ee6a74d16