Always wrap the entire test command with unshare
Summary:
In this diff we wrap the entire command to execute a test with `unshare` when the `--experimental-force-linux-network-sandboxing` is provided. This means that all other wrappers that are potentially added around the test binary are also wrapped in unshare.

## Why is wrapping wrappers needed?

The llvm coverage wrapper that we use to collect coverage for cpp unittest
binaries assumes that the first argument that it receives is the test binary to
execute.

It then uses the path of such binary to locate shared libraries and extract
coverage information.

This assumption is broken by the new tpx flag
`--experimental-force-linux-network-sandboxing` which wraps test binaries by
prefixing them with `unshare --map-root-user --net`. As a result, the llvm
coverage wrapper infers that `unshare` is the test binary and cannot locate
shared libraries.

IMPORTANT: We need to build the llvm coverage wrapper from source because otherwise this is served via msdk which will fail to download the wrapper when wrapped in an `unshare` call.

Reviewed By: asm89

Differential Revision: D69180841

fbshipit-source-id: ffaa72760070072d5e215f0b0ae391f940784f32