Bump Min ARVR Version for oculus/kiosk:kiosk aidl to 35
Summary:
With the SDK35 upgrade, some aidl files require a new version of the SDK in order to parse annotations and build correctly. Unfortunately ARVR cannot be fully bumped to 35 or 36 due to issues with the NDK.

To be compatible with this bump, we need at least SDK35, so I'm bumping some of the affected files to SDK-35 explicitly.

If these are your files and you think there's a business reason to stay on the current SDK, please let me know so that we can work on finding a solution. Otherwise, this is likely a very safe change.

Differential Revision: D69413924

fbshipit-source-id: 22aaab4ae88a588f050c3fd8c63fe06a14bcc7c7