support `--overall-timeout` in `buck2 build`
Summary:
Like it says in the title. I would like us to make progress on making Citadel
timeouts actually work, and to do that we are inevitably going to need to be
able to enforce timeouts on `buck2 build`.

This adds that feature. It doesn't aim to cover everything (yet), such as
listing or analysis, but it should offer a decent place to start.

Reviewed By: JakobDegen

Differential Revision: D69401528

fbshipit-source-id: 9f415953a4979a59d6d976e4bdb7bcbf2bf1764d