Allow users of `system_cxx_toolchain` to specify the archiver (#855)
Summary:
This is useful for cases such as when working under a conda environment and you intend to use the conda toolchain exclusively without requiring a toolchain installed on the system otherwise. For whatever reason 'ar' is only provided with the full toolchain prefix.

Likewise this would let you specify `llvm-ar` for the same reasons.

Pull Request resolved: https://github.com/facebook/buck2/pull/855

Reviewed By: ezgicicek

Differential Revision: D69300208

Pulled By: IanChilds

fbshipit-source-id: c000e8057912ac7722df4fda7c90d5e7a69a582f