Add health check context to maintain state
Summary:
This is a per command context that will allow us to cache some state information to be used across the health checks.

Context:
For the stable revision check, we want to warn the user if they are seeing excess cache misses on Build/Run/Test commands. For this, we need 3 pieces of information:
1. Mergebase rev hash: From VersionControlRevision event.
2. Has excess cache misses: From the ActionExecutionEnd event.
3. Command: From CommandStart.

Intermediate:
Before we integrate the stable revision check with the new design, we will source this data from the simpleconsole::event_observer.

Final state:
The HealthCheckClient will become an event listener running in the buck2 client space. The client will parse a subset of the action and snapshot events to cache the relevant information and send a subset of the information to the HealthCheckServer via gRPC.

The goal of this context is to reduce the grpc overhead.

Alternatives considered:
1. Pass the events to the health check doctor. However, this would be prohibitively expensive since the `excess cache miss` would need all of the action events which are way too many compared to Snapshot.

Reviewed By: IanChilds

Differential Revision: D69521829

fbshipit-source-id: 74b667c9842093bdfcebc472a86ba8a3fcc2772d