Record previous command with mismatched configs
Summary:
**Context**: A common source of build speed regressions is commands that come in between user invocations that invalidate buck's state without user's knowledge (caused sevs like S443499). Currently, we have no way of diagnosing these. To alleviate this issue, we would like to keep track of such previous commands that have invalidated buck state.

As a first approximating for that, we are checking if external configs have changed between the previous command and the current one. If so, we fire an event `PreviousCommandWithMismatchedConfig` that keeps track of the previous command with mismatched config ( `uuid` and `sanitized_argv`).

**Q: Why just external configs?**
We don't do this for all configs (i.e. including the values of project-relative configs) for now but only for external configs because they are easier to diff against. In the next diffs, the plan is to add diffs between the current command and the previous invalidating command so that people can make more sense of it.

Reviewed By: cjhopman

Differential Revision: D69185805

fbshipit-source-id: 922e336ba015bafe4a78ddbfeed9302f4e61287f