Decouple EventsCtx from BuckdClient
Summary:
Currently EventsCtx is owned by BuckdClient/FlushingBuckdClient, which makes it awkward to log events before and after a daemon connection is active.
The main goal of this is to make EventsCtx available after a command has failed so we can finalize logging errors in one place and propagate them to the command report.

The trade off is needing to manually pass an EventsCtx arg for each streaming command, but this should also allow cleaning this code up later since it should also make it easier to share more logging set up code between daemon and non-daemon commands.

All the important changes are in `buck2_client_ctx`, the rest is just changing commands to pass `events_ctx`.

Reviewed By: iguridi

Differential Revision: D69164808

fbshipit-source-id: 13ce78e2ee3efbbf24c5704037e6c62feb236b8b