Finalize subscribers instead of depending on drop and AsyncCleanupContext
Summary:
This is the other main piece to achieve the goals in D69141560.

By calling `finalize` we can pass any errors directly to the command report that may have occurred when dropping subscribers.

This could also allow make it possible to add dependencies between subscriber finalization later on, to write the invocation record to the event log, for example.

Reviewed By: iguridi

Differential Revision: D69217731

fbshipit-source-id: d978ec3d2d508a860fa8c1e1c9daea8a45e6359c