Make AsyncCleanupContext cleanup explicit
Summary:
This stack has 2 goals:
- Propagate errors that happen during log upload to the command report to be uploaded to buck_results.
- Reduce the number of places that errors can 'leak' when processing events and uploading logs.

`AsyncCleanupContext` is used to register async code to be executed on drop, which is used by some `EventSubscriber`s to upload logs at the end of the build.
In order to propagate errors from uploading to the command report we could move command report generation into `AsyncCleanupContext`, however this would still leave places where errors could 'leak' because we rely on a specific drop order and each subscriber successfully registering to be cleaned up, which seems fragile.

Doing this explicitly makes this more predictable, makes it easier to propagate errors to the command report and simplifies refactoring in this stack.

In theory the current design does mean any EventSubscribers will get cleaned up automatically no matter where they are created, which is no longer true. We create and drop subscribers that use cleanup context in 2 places, in `exec_async` for streaming commands, and in `instant_command`, so this means a small amount of duplication, but I plan to merge these into one place in a later diff.

Reviewed By: iguridi

Differential Revision: D69141560

fbshipit-source-id: db32d2fb2f091d57c512b287cae41f7329057728