EdenConnectionManager: provide default semaphore
Summary:
# This diff

Alters the EdenConnectionManager constructor to take an optional semaphore value, allowing the user to trust the default value that is selected by the constructor.

# Context

In most cases, the client doesn't actually need to control the semaphore value used to limit the number of concurrent Eden thrift requests. Currently, all clients use the same value (2048) but they all specify this value separately. This makes it difficult to make blanket changes to the value because we have to change 4-5 different harcoded values.

In addition, even if the client selects a value, it may not be the optimal value to keep Eden from issuing Queue Timeouts or throwing "OS Error 60 (timeout)" errors for filesystem IO.

Therefore, this diff sets the default value in the Eden client library and assumes that all current clients want to use that one instead of their current hardcoded default.

Reviewed By: JakobDegen

Differential Revision: D69430589

fbshipit-source-id: bfa16dee0d5b8e515fb40f3750cdb5adaa26f166