add EdenFS health check
Summary:
# This diff

Adds logic to daemon startup to register a periodic task that checks the health of the Eden daemon. This periodic task is limited to fbcode builds, and it should only get registered when the repo root is part of an EdenFS checkout.

# Context

We'd really like to understand how often Buck2 observes hanging EdenFS mounts in the wild. This will help us determine how often users are seeing hanging mounts, and also whether current mitigations we're adding to EdenFS will decrease the frequency of hanging mounts.

The best way to do this is to add a periodic task that's registered at Buck2 daemon startup time. This task will run every X minutes (currently 10), start a new connection to the Eden daemon, and issue a single getDaemonInfo() command. If getDaemonInfo() returns a successful result, we can be sure that the Eden daemon is not hanging. If it returns variations of some errors, we can flag a potential hang.

# Technical Details

Buck has a library which allows the Buck daemon to easily connect to the Eden daemon and issue thrift requests. Each Thrift request returns a Result which can be a ConnectAndRequestError. There are two variants of ConnectAndRequestError.

* The first is ConnectError, which indicates we could not establish a connection to the Eden daemon.
    * This error is only indicative of a hang if it was caused by failed IO into the Eden mount.
* The second is RequestError, which comes in 3 flavors.
    * ThriftError, which is never indicative of a hanging mount (our assumption is that it's caused by a bug in thrift code, not an error from interacting with the Eden daemon).
    * ApplicationException, which is indicative of a hanging mount if it has the flavor 'Timeout' or 'LoadShedding', which means the Eden daemon was unresponsive or slow to respond.
    * Ex, which is a generic exception. We ignore these errors for now.

A periodic task issues getDaemonInfo requests to the daemon which have a chance to be one of the above errors. For each getDaemonInfo response, we check if the response is an error, and if so, we check if the flavor of the response (see above list) is indicative of a hanging mount.

Reviewed By: rajneesh

Differential Revision: D69430945

fbshipit-source-id: 589ccbba9eb0e29580bf857cfbfe362d1fb37fcb