add option to specify an alternate application name
Summary:
We currently hard-code the application name to be equivalent to the target name. However, when generating thrift targets, we can end up in the situation, where we will have duplicated application names in a dependency tree.

This option allows us to generate unique application names, either based on user overwrites, or based on the namespace.

Reviewed By: michalmuskala

Differential Revision: D69687014

fbshipit-source-id: a2b95992fd021e4b6baaabb555e506b9f05744ae