add safety checks to prevent two applications in the same dependency tree
Summary:
We currently can have two application with the same name in one dependency tree. Currently, one of the application will overwrite the other one, which causes incorrect build or build failure, and undefined runtime behaviour.

This diff makes the build fail when there are two applications with the same name in on e dependency tree.

Reviewed By: michalmuskala

Differential Revision: D69662053

fbshipit-source-id: e39affd22ddd2ed7341179ffe863f5647f61f850