Turn off the low-pass filter for cxx_precompile
Summary: At the suggestion of the Buck team, to allow hybrid/local execution when congestion is high. This shaves off a good 30 cxx_precompile actions or so from the beginning of the critical path in full cold builds of current modular builds (e.g. ipnext_client) and saves about 1min 30s.

Reviewed By: IanChilds

Differential Revision: D69755991

fbshipit-source-id: 4793664444629f4e8042f3a6d11187c702dcce51