lifetimes: Fix unsound APIs on `ProviderCollection`
Summary:
The rough shape of these APIs was `&'v FrozenProviderCollection -> &'static SpecificFrozenProvider`. This was "justified" under the knowledge that the frozen heap in which the provider collection was allocated kept the individual providers alive too.

There isn't really any agreed upon standard as to what the contracts regarding sound lifetime handling around frozen values are, but I think such an API has to be wrong - the lifetime of frozen heaps is not, in fact, `'static` and APIs that pretend it is are wrong.

Reviewed By: IanChilds

Differential Revision: D69713871

fbshipit-source-id: 5d9f894e32e3451b815acfff38c0cc21a8bcbea2