Rename 'materialize_build_outputs_from_previous_run'
Summary:
If I understood the code correctly, `fn materialize_build_outputs_from_previous_run` materializes the build outputs at a given path. The only callsite for the function currently checks to make sure the previous action wasn't cleaned up, and thus the build output for the previous run would be materialized, but the function itself does not necessarily materialize the build outputs from the previous run.

I've updated the function name and comments to reflect this as it was confusing to me when I was first going through the code

Reviewed By: JakobDegen

Differential Revision: D69562568

fbshipit-source-id: 9224aa0f7f9dd10fc5ee7f2e56bea31683ba856f