avoid re-evaluating to know if a package has changed (#859)
Summary:

Context: Adding a random file to `fbcode/` does some recomputation in the graph that leads to 40s build time in the incremental case.

When `fbcode/BUCK` is a dep, this is unavoidable since you need to traverse the entire graph to understand if you need to recompute anything. However, in the case of some targets like `//aps_models/ads/gmp:launcher_with_publish` `fbcode/BUCK` is not a dep, but since we check for nearby package files for the entire graph, we end up spending more time.

#thanks jakobdegen for helping me figure out what causes additional 35s computation in dice when we add files to fbcode/

Reviewed By: JakobDegen

Differential Revision: D69714496