Read cxxflags from cxx_toolchain instead of buckconfig
Summary:
The cxxflags et al. are read from Cxx Toolchain instead of from the buckconfig. Moving forward these flags will no longer be in the config and only in the Cxx Toolchain (T213442427).  There is no observable behavior change.

The main motivation behind (T213442427) is to improve buck-build-graph caching between dev and opt build-modes by moving the cxx toolchain out of config files and directly into the toolchain target.

The cxxflags on the `CxxToolchainInfo` need to be indirectly read with `actions.dynamic_output` because they are type `cmd_args`.

### flags containing string parameter macros
I do nothing about string parameter macros (e.g., `$(location ...)`) in flags  because 1/ I do not see string parameter macros in the flags of the CxxToolchain, and 2/ the ambiguity of how to handle them.  They will need to be handled if/when they become a problem. I mention the string parameter macros because before this change, when the flags were read from the buckconfig, the code would filter out strings with `$(` and `)`.

### CxxToolchain Info
The CxxToolchainInfo is loaded from `toolchains//:cxx` that is configured to the platform `ovr_config//platform/linux:platform010-clang` based off where the flags are currently read from (`fbcode-platform-cxx#platform010-clang`).

Reviewed By: scottcao

Differential Revision: D69617220

fbshipit-source-id: 958f94cb4674d6904432d594ec5938c64dd1067d