fs_util: refactor retry_io test
Summary:
# This diff

Refactors fs_util retry_io test to make it easier to add to in the future.

# Context

In the next diff, I'm going to add a new IO error type that should be retried. Adding a test for the new error type was painful, so I refactored the code to make it much easier in the future. Now we can easily add to the list of test cases when new retries are added.

Reviewed By: JakobDegen

Differential Revision: D69890300

fbshipit-source-id: cfcd1480aae9b86b6a1132cf8a63558530e3ff9f