Support passing extra linker flags to the relinker
Summary: It's possible for a build to only want a specific optimization during the final phase of linking (aka relinker stage). An example would be to only run LTO during the relinker stage since LTO can incur a high build speed cost. We've tested internally that this doesn't affect APK size is neutral, but we'll still need to run perf tests. Support this flag as a general case so we can also run other specific optimizations only during relinker stage.

Reviewed By: smeenai

Differential Revision: D69625538

fbshipit-source-id: e934b6e076853f5b3593d704d09d539355e20ebc