Keep track of project-root's .buckconfig.local values in event log
Summary:
Currently,  values of project-relative buckconfigs are

1) not tracked in event logs
2) not taken into account when detecting previous command with mismatched configs (D69251704).

The reason is that they are huge. It is unlikely that devs would change project-relative buckconfigs but one exception is project-relative `.buckconfig.local` files which might be changed.

For reproducibility and correctness, we would like to store them in the event logs and count them as part of mismatched-config tracking of previous commands.

We do this by simply parsing values of project-relative .buckconfig.local in `get_buckconfig_components` which is used for both 1) and 2). The order is kept consistent so that first we have external configs, then .buckconfig.local and then the commandline args.

Reviewed By: scottcao

Differential Revision: D69860771

fbshipit-source-id: 1f2f8efc8104ab1b084c124c52f5465d6a4ae175