process all sapling status cases
Summary: Process all sapling status operations. It's not necessary to do as we call sapling with `-mardu` params which stands for modified, added, removed, deleted, or unknown, so other operations should not be reported. But I want to keep code consistent with `edenfs-client` D70003870 as it would be good to use the lib in the future.

Reviewed By: rajneesh

Differential Revision: D70092436

fbshipit-source-id: 0afa81e88943c058b4fbfe2e62cdc22d9e427414