configuration: Full test suite for evaluation order in configuration
Summary:
The main configuration logic has this thing where, in general, all parts of it can exhibit two possible semantically distinct abnormal conditions: Incompatibilities and errors. Both of these are implemented via short circuiting. Unfortunately, that means that the semantics of when exactly incompatibilites happen and when errors happen depend on the exact order in which different things are evaluated. The most egregious cases of this were somewhat addressed in D46910954 , but there's still a lot more there.

Besides that, the way that configuration and platform deps are written is also super weird and causes surprises like things being evaluated in dead select branches.

This diff adds a set of tests that attempt to precisely reproduce all of the existing observable behaviors, replacing a couple existing tests in this space. This is not an endorsement of those behaviors - some of them are clearly ridiculous - but it's a necessary step for anyone to be able to understand the impact of their changes to this code.

Reviewed By: IanChilds

Differential Revision: D69713869

fbshipit-source-id: b21497388f81df5ee89485a9396061589f397804