rust-project: Include buckified sysroot in project
Summary:
rust-analyzer wants to drop support for their "stitched sysroot" (where it will assume the structure of the sysroot) in favor of relying on the build tool to tell it the sysroot layout. Since we have a buckified sysroot, that can be buck, and we can include a nested JsonProject for the sysroot.

Corresponding rust-analyzer change: https://github.com/rust-lang/rust-analyzer/pull/19096

Reviewed By: davidbarsky

Differential Revision: D69328620

fbshipit-source-id: ee0b461b2e144bfefd30970d6f6ce35ae89e8ee1