configuration: Add `ConfigurationDepKind` to coerced attr traversal
Summary: This changes the way traversals work on coerced attrs to just have a single `configuration_dep` function and pass to that function an enum indicating where the configuration dep appeared. This makes it possible to handle these things in a more uniform way. Will be used in one of the next diffs

Reviewed By: IanChilds

Differential Revision: D69713858

fbshipit-source-id: a41c9a5ca6df409d252fe9cf2eda03a076b3a73f