Add `StrongHash` trait
Summary:
Add a `StrongHash` trait. This is similar to `std::hash::Hash` trait except the digest produced is a `Vec<u8>` that can accommodate hashes of any length. Also implemented is a proc macro for `StrongHash` such that it's possible to derive `StrongHash` for trivial implementations.

I will use this later to obtain a cryptographic hash for configuration in D64873220 to fix our existing problems with hash collisions. This can also be used to make equality better for `ConfiguredTargetNode`.

I decided to add `strong_hash` and `strong_hash_derive` to gazebo because these crate seem pretty foundational and I need to add a dependency from `starlark_map` later in the next diff.

Reviewed By: JakobDegen

Differential Revision: D64851950

fbshipit-source-id: aaf83289c3c24857f3271d21a3ab9b7aab3fd53e