Use truncated blake3 for configuration hash via StrongHash trait
Summary:
Use truncated blake3 for configuration hash instead of the default hash via `StrongHash` trait. This will make our configuration hashes significantly less collision prone.

Note this change will update all buck-out paths and thus action keys. In order to avoid changing everything as part of the version bump, this change is gated by a buckconfig so it can only be enabled by a buckconfig. The buckconfig can also do a percentage-based rollout based on the new configuration hash, so it can be rolled out via a deterministic incremental rollout if need be.

Reviewed By: JakobDegen

Differential Revision: D64873220

fbshipit-source-id: 608e0804ca2b9748c057074eae30dc484664918d