Implement testing allowlist
Summary:
For sake of preventing arbitrary usage of CLI modifiers/modefiles containing modifiers, we have an explicit allowlist for CLI modifiers can be used on CI. However, this can only be entirely enabled or disabled, and there are several scenarios that would benefit from more fine-grained allowlisting. One scenario is e2e testing. It's perfectly valid to use most cli modifiers for e2e tests, but those cli modifiers often don't need to be allowlisted for use cases besides testing.

This diff implements a separate allowlist for CLI modifiers on CI when used as part of e2e tests. This only allowlists modifiers for constraints in the list when running under e2e test. We check for whether we are in an e2e test environment by reading the `buildfile.extra_for_test` buckconfig.

Reviewed By: samkevich

Differential Revision: D69970561

fbshipit-source-id: fc1ed7b117763489e18ec1f059c9e97d02405948