Move error sink to OutputStream
Summary:
"Originally, both the output sink and error sink were stored in a separate `OutputStream`. However, for the error sink, the other fields in `OutputStream` are unnecessary.

In this diff, I added a new field called `error_sink` to store the error output sink. This change eliminates the need to create a separate `OutputStream`. `OutputStream` now has a stdout sink and error sink.

Reviewed By: IanChilds

Differential Revision: D70367972

fbshipit-source-id: ecf280a1e39d2479e5ba157b29b9cd79bcf5220b