transitions: Implement new transitions API
Summary:
This diff implements the API proposed [here](https://fburl.com/gdoc/xf4jeu6n).

The main structure of this change is that the `TransitionId` type is turned into an enum representing the old and new kinds of transitions. The code in `buck2_transition` is then updated to handle both branches. Additionally, `attrs.transition_dep()` is updated to accept a target label for `cfg`. This target is reported as a configuration dep of the main target. Incoming edge transition support follows in a future diff.

Reviewed By: cjhopman

Differential Revision: D69798612

fbshipit-source-id: d01967a5eba5b2b6f94eec3a4d813eaad251cecf