transition: Implement incoming transitions in new API
Summary:
Mainly, this modifies the `cfg` value in the `Rule` object and then fixes up the places that create it/use it.

I think actually the way I just changed the internal attrs wasn't the best I could've done, going with types instead of consts would've been a bit cleaner. Oh well, this is still an improvement and mostly ok as-is

Reviewed By: cjhopman

Differential Revision: D69798607

fbshipit-source-id: c38072a11a4705309b8059e70293f1ce28c2dd07