run: Less verbose "we're running your thing" message
Summary:
This is an alternative to D69959932 . I like that idea, but it's just a bit too much on console I think. Consider this example that I ran into:

{F1975491856}

Putting aside all the crap that buck leaves behind on stderr, this would normally have been 5 lines of very high signal stderr output from the binary being run - the additional two lines added as a part of the run output message hurt the readability of that significantly, and that seems a bit much.

I'm hopeful that the explicit "running your binary" achieves some of the goals of reducing user confusion about it being buck's fault when that binary crashes. We might have to find some more in-between options, though.

Reviewed By: IanChilds

Differential Revision: D70275291

fbshipit-source-id: 893f59ef8dbc994e18f67d280f2e0f3b36001add