Fix poorly isolated tests
Summary:
We use a `filegroup` target to copy things in some "isolated" tests, which is now broken because this file is behind a subpackage and isn't seen by the filegroup. Fix this temporarily.

The shittily isolated tests absolutely must go, but in the meantime, this is the best we have

Reviewed By: dtolnay

Differential Revision: D70468549

fbshipit-source-id: 6024f61b0023df41d3491e87bebfc6621250c6e7