Record whether command should restart in invocation record
Summary:
Currently we exclude failed commands from UBE-0 if they restarted and the restart succeeded by doing a self join on restarted_trace_id, but we can't do this in scuba and it's slow.

Excluding commands where should_restart=true doesn't guarantee that the restart was successfully triggered and that the restart logged a result, but it should be a close approximation.

Reviewed By: Will-MingLun-Li

Differential Revision: D70509987

fbshipit-source-id: 37562244b61c99b7f8dedd5ce91064acccf7ba17