Update 'materialize_failed_outputs' into a proper option
Summary:
Update `materialize_failed_outputs` as a proper build flag. Enabling this flag will materialize all failed action outputs that are declared which could be useful for debugging purposes.

There's no explicit request for it, so I don't have a strong opinion on this, so if someone has a strong opinion about not keeping this flag that'd be fine with me too. Just thought I'd add this since the work is done anyways.

Reviewed By: IanChilds

Differential Revision: D70206452

fbshipit-source-id: 8511e2c532c39139c479f15d2a5f79495fd9903a