Disambiguate fs2 extension-trait locking calls
Summary:
Some `fs2::FileExt` extension-trait methods overlap with new file locking API on `std::fs::File` in Rust 1.84 (D70601924).

```lang=text,counterexample
error: a method with this name may be added to the standard library in the future
   --> fbcode/clifoundation/cli_target/validator/src/main.rs:279:20
    |
279 |         cache_file.lock_shared()?;
    |                    ^^^^^^^^^^^
    |
    = warning: once this associated item is added to the standard library, the ambiguity may cause an error or change in behavior!
    = note: for more information, see issue #48919 <https://github.com/rust-lang/rust/issues/48919>
    = help: call with fully qualified syntax `fs4::FileExt::lock_shared(...)` to keep using the current method
    = note: requested on the command line with `-D unstable-name-collisions`
help: add `#![feature(file_lock)]` to the crate attributes to enable `std::fs::File::lock_shared`
    |
1   + #![feature(file_lock)]
    |
```

This diff (still on Rust 1.83) qualifies all the extension-trait calls so that these do not become a factor in the Rust 1.84 diff.

After D70601924, it would be reasonable to drop `fs2` and `fs4` dependencies and use the standard library APIs instead.

Reviewed By: diliop

Differential Revision: D70676686

fbshipit-source-id: ed9dc8fa139c56e1a6f8e2fc39b923a81e8711b3