leverage CXX library for FFI
Summary:
# This diff

Adds CXX support to the system fingerprint library so that C++ code can leverage the system fingerprint utility.

# Context

The EdenFS daemon is written in C++. The Eden team would like to include system fingerprints with all telemetry samples, but that requires the Eden daemon to call into Rust to collect the system fingerprint.

We can easily enable this by leveraging cxx and defining a `fingerprint()` FFI function for obtaining the system fingerprint from C++.

Reviewed By: gmambro

Differential Revision: D70652112

fbshipit-source-id: debeea6926b2641d0786784a50d48f00a5605ad9