Merge unspecified and generic error classification
Summary:
To simplify configuration and since there are a few more unspecified tags that should also be excluded from category keys as generic errors.

ClientGrpc is a generic error but not an unspecified error so making it unspecified, all tags assigned to ClientGrpc errors should be tagged infra, or at least, none should be tagged as input.

Reviewed By: Will-MingLun-Li

Differential Revision: D70733102

fbshipit-source-id: 01c3f0b6f2115e138557071fe5574f65d0e0d5d8