Improve UX when swift-ide-test has an error
Summary:
Currently a failure of `swift-ide-test` just prints a Python exception with
a very long array with the failed command line, but fails to communicate what
actually failed, which might be confusing to the user.

After the changes, if swift-ide-test fails, it will print the stderr of
swift-ide-test and exit with the same return code (so the process still fails).

Reviewed By: ellishg

Differential Revision: D70433723

fbshipit-source-id: f1c47df16b3169afca31aba0bdaa9bc9b716c132