Remove unused parts of Darwin dist_lto_prepare script
Summary: Distributed thin-lto requires you to pass object files to the linker directly, surrounding them in `--start-lib` `--end-lib` flags if you want archive semantics. This script is used to take static archives and unarchive them such that the contents can then be passed to a linker invocation. It is meant to support both thin archives and regular archives, but in my estimation, it doesn't actually work for thin-archives. For the time being, we won't be needing thin archives support. Any archive we build ourselves we will just pass the object files directly instead of putting them in an archive, then extracting them again. The only archives we still have to extract are vendor provided archives (ie. from the platform SDK), and those up until now are conventional static archives.

Reviewed By: rmaz

Differential Revision: D70174175

fbshipit-source-id: 7d2866e859719d3d1f67ad4bc1096bda786a4831