Prefer files by the current user
Summary:
Previously `rsync` would try to preserve the owner (`-a` is the same as `-rlptgoD`, and `-o` sets the owner) when merging files in index stores. This is problematic if we copy a file owned by root, and then try to update it.

Instead, ensure the merged output file is always owned by the current user. This is `fbnobody` when jobs are run on RE.

Ideally the index store would never contain files owned by root, but this prevents problems when merging or using the index store.

Whilst we're here, clarify the comments in the merge tool to explain why we use rsync.

Differential Revision: D70811661

fbshipit-source-id: fe836147793e875d6bf10c51a9e79ee6f2637d53