Add tags channel writer to client
Summary:
The health check client needs to pass 2 types of reports:
1. A warning/notice to display on the console.
2. Information to be recorded in scuba.

The solution is to create 2 channels and have the health_check_client write these reports to the channels and on the other end, read these reports in simple/superconsole and recorder.

This diff stack adds the scuba logs. Currently the scuba logs are only tags and hence the health_check_client will write a Vec<String> to represent the tags. State is maintained at the recorder and the health_check_client remains stateless about the reports.

Working prototype implementation: D70638457

Reviewed By: ezgicicek

Differential Revision: D70638282

fbshipit-source-id: d70a459e65a1a251b20f5e676341de644208a6aa