Reproduce inconsistency/bug in `wait_for_idle`
Summary:
The docs on `wait_for_idle` and `is_idle` are a bit inconsistent in what "idle" supposedly means:

```
    /// Wait until all active versions have exited.
    pub fn wait_for_idle(&self) ->

    /// true when there are no tasks pending cancellation
    pub async fn is_idle(&self) -> bool {
```

The behavior that's implemented is what is implied by `is_idle`, ie this only checks if there are tasks still actively being cancelled.

Tbh, that seems like the wrong behavior.

As far as I can tell, this does not create a bug in buck2 - from reading the code it does seem like we'll always drop all outstanding `DiceTransactions` strictly before relying on this, so it should be ok

Reviewed By: cjhopman

Differential Revision: D70851255

fbshipit-source-id: b6ff861a64eb6f994fb865409aed37f794989045