Implement dict of artifact and artifact value to be exposed through error handler
Summary:
Get the list of `OutputArtifact` provided and convert it into a `SmallMap<StarlarkArtifact, StarlarkArtifactValue>` to be exposed for rule authors to work with. This allows for easy access to artifact values with the artifact provided.

Also check if the requested output actually exists at this point, and return an error if it doesn't. This way both local and remote executed actions will have the error handled if the file is missing

Reviewed By: JakobDegen

Differential Revision: D70648771

fbshipit-source-id: 9a5990c0562e491d2960c280ed1d9eb6f141823c