Add 'outputs_for_error_handler' parameter to run actions
Summary: Extending Run Action API to take in an `Artifact` value to be materialized in the event of a failure. This will be passed into Action Error Handler and wille be useful for languages that generate an error diag file, additional implementations will be added to ensure the file can then be parsed/processed and passed to error handler for structured action error purposes.

Reviewed By: JakobDegen

Differential Revision: D70206451

fbshipit-source-id: d74fd369b34a581ada8a6abb0ba8250a17912990