Add tests for error handler with output artifacts
Summary:
Add tests for extended error handler API that can take in an output artifact and provide that to be used to provide action error diagnostic information.

Tests cover reading both string and json for actions executed both remotely and locally. Also covers the scenario when the provided output artifact is not found, which will display an error message

Reviewed By: JakobDegen

Differential Revision: D70660468

fbshipit-source-id: ea3b607a67808b080681559777c21089a759ecfe