Always return an artifact
Summary:
Previously we wouldn't call `merge_index_store.py` when we had zero index stores to merge. As of D70747622 we can call merge_index_store.py with zero inputs (it receives an empty argsfile and invokes rsync zero times).

This makes the surrounding code simpler, and simplifies later diffs in the stack.

Differential Revision: D70530009

fbshipit-source-id: ca787d85fa106aaab7465f5e5c648a0233a097de