Add timeout for awaitTermination
Summary:
On windows, `grpcServer.shutdown().awaitTermination()` seems to be blocking indefinitely for some reason.
Adding a timeout so the installer shuts down and we don't end up with zombie installer processes, since those lead to build failures on windows if the installer is rebuilt due to rebase or config changes (`The process cannot access the file because it is being used by another process. (os error 32)`)

Reviewed By: navidqar

Differential Revision: D70910672

fbshipit-source-id: f748b35df5c1d4b990ccdd79bbeadfe665925e9d