buck: re: Actually set `remote_execution_dynamic_image` if specified
Summary:
`remote_execution_dynamic_image` was introduced to `CommandExecutorConfig` last year, but I had to stub this out in D67774608 because some asic usages had a pinned buck2 version (S473397). I didn't conditionally set this attribute correctly though - I should have used `hasattr` instead of a `dir()` on the type, so we never actually set this.

We only discovered this because on D68419981 we didn't see the custom image being used when we expected to.

Reviewed By: wenxin0319

Differential Revision: D70997480

fbshipit-source-id: 18de050c8874ee879e9d0ef66ae4ba1ed6aed95f