Print ActionErrorDiag message in superconsole
Summary:
ActionErrorDiag messages are currently being written to simple stderr, so we can see it in simple console and Buck2 UI. However, it was never written to superconsole.

Updating the behavior so it will be printed in superconsole only if there's an additional message associated. Since the additional message is usually expected to be a wiki or some sort of hint/instruction, gave the message some color to be more obvious. (Chose DarkCyan as I thought it looked the best, happy to change to something else if there's a general convention)

Reviewed By: ezgicicek

Differential Revision: D70795480

fbshipit-source-id: deb9c87eb2ae32872018ec1f9d566daf3dde58fe