Add error handler for CXX
Summary: Adding an error handler for cxx, only handle duplicate symbol errors for now but can easily be expanded in the future. Motivation is that in Buck2 support groups, we get lots of questions regarding `duplicate symbol` errors. Adding an error handler to categorize the error but more importantly direct users to a wiki to better understand the error

Reviewed By: ezgicicek

Differential Revision: D70795708

fbshipit-source-id: 5a6bdf963a1ee023202847170a1521e055f9c3ae