wrap a type later
Summary:
Like it says in the title. This makes my next diff more clearly correct since I can now avoid matching on a specific enum variant and erroring out at runtime if it's not correct.

This does mean we've got to do a copy at some point, but this definitely does not matter.

Reviewed By: blackm00n

Differential Revision: D70893980

fbshipit-source-id: 9bcd3d894511e821c2b4d88b64d932eb231e7f74