Revert D70913453 --force-non-staged with InstrumentationTestRunner pm install for APEX
Summary:
In D70913453 we added --force-non-staged to make sure the APEX was immediately applied after installation. However this didn't restart vrruntimeservice and so tests expecting a sideload would still get the old vrruntimeservice and not the new one. The only way to make this happen is to restart the service or to reboot.

In D70938892 we implement a pre_test.sh that adds an adb reboot, solving the original problem that D70913453 tried to fix. Now that we do the reboot, we can revert the original change since it is no longer needed and potentially added other unknown side effects.

Reviewed By: vener91

Differential Revision: D71050654

fbshipit-source-id: 8524eeb3be789f73128cf62bbf65ed0830faae59