Retain logs from previous daemon under daemon-dir/prev
Summary:
To debug cases of the daemon being killed and unexpectedly needing to restart.

Keeping the entire directory and not just buckd.stderr because if the daemon shuts down with a structured error the error should be in buckd.error.log but not buckd.stderr.

Logging the reason for daemon shutdown (if it's not a crash) to stderr is addressed in the next diff.

Reviewed By: iguridi

Differential Revision: D71017221

fbshipit-source-id: baf544ed8e0e0558a7075746e1a269f6d2f0b6ba