Refactor the code to return buck2_error::Result first
Summary:
In starlark api definition, we need to return `starlark::Result`. For some error, we cannot directly convert the error to `starlark::error::Error`, we need to convert to `buck2_error::Error` first. As a result, we can not use `?` syntax in there.
In code, we have such lines
```
 if let Err(e) = writeln!(this.sink.borrow_mut()) {
            return Err(buck2_error::Error::from(e).into());
        }
```
Also I will help for  new apis `ctx.output.stream/stream_json` in later diff because these will have similar logic with `ctx.output.print/print_json`.

So this diff, we implement a method to `OutputStream` to return `buck2_error::Result` first and call this method in starlark api definition.

Reviewed By: IanChilds

Differential Revision: D70418114

fbshipit-source-id: 02982ce59621dc76a94e7ce4d49b53e0a686aa17