Add display to superconsole
Summary:
Read from the channel and display warnings on superconsole. This caches a list of `active_warnings` when the channel has some data.

If the channel is empty, it just means that the health check client hasn't published anything on it and we should continue to show the existing warnings.
When the health check client completes its run_checks, it will publish a DisplayReport for each of the health checks. This will ensure that we are always communicating the entire state back to the consoles.

Reviewed By: samkevich

Differential Revision: D70980867

fbshipit-source-id: 5cb0720f061ac6d50e2dee2e6c293dd7b6bd14bd