Add DisplayReport channel
Summary:
Add the channel reader and writer to send reports to be displayed on the console. In this diff, it is only propagated to the `get_console_with_root` since adding the field to the simpleconsole affects a lot of files and makes it harder to review.

Prototype of all changes: D70638457

Reviewed By: samkevich

Differential Revision: D70980858

fbshipit-source-id: ed91b7e6e462c8f24b5c9c2f4f49c9db0a3cb98e