Fix test_apple_coverage_xplat test
Summary:
A previous fixed the llvm-profdata and llvm-cov target executable to be Pika consistently. Specifically the test was breaking when the test ran on a mac as a result of the test using the system llvm-profdata llvm-cov tools.
This is a slightly different issue: it's not the LLVM tool, but how the test bundle is built. Because it's in xplat, xbat is the default toolchain.
This changes the build to force pika.

Not sure why this wasn't surfaced while testing the bump, and is failing only after the bump landed.

Reviewed By: IanChilds

Differential Revision: D71165372

fbshipit-source-id: c9124befcffc55ef4acac44aff44c08d590aa7c7