log sandcastle type to metadata
Summary:
Like it says in the title. This is helpful to filter down to builds that are
runnign on a specific SC type, which is often the granularity at which we
experience queueing.

This is set by Sandcastle: D47606242

Putting this in `ci_identifiers` felt a tiny bit weird but it's not really more
weird than having the alias here (it's simply even less specific), so it seems
coherent.

Reviewed By: samkevich

Differential Revision: D71310658

fbshipit-source-id: ca949359981035fbea9f85ab9632add4320deea9