Add DepOnlyIncompatibleInfo provider
Summary:
Context: We are migrating to a stricter compatible with behavior (known as `error_on_dep_only_incompatible` or safer `compatible_with`) due to prevent targets from losing CI coverage without notice. This rollout is currently done on a target pattern basis. Although we have ways to automatically surface pre-existing errors as soft errors, it's still a bit difficult to know whether this is safe to enable for a set of targets, leading to cases where we thought it would be safe to enable but that led to breakages and reverts.

This and the next diffs try to fix this by introducing a way to surface soft errors for `dep_only_incompatible`. Instead of using one category for all errors, we can make buck2 populate user-defined soft error categories for custom target patterns, making it much simpler to know whether `error_on_dep_only_incompatible` is safe to enable for a set of target patterns.

The setup for this is that there's a target producing a custom provider `DepOnlyIncompatibleInfo`, which contains a map of custom soft error category to a list of target patterns to check. Then we set a buckconfig `buck2.dep_only_incompatible_info` to point to this target for buck2 to use this provider.

This diff sets up the `DepOnlyIncompatibleInfo` provider.

Reviewed By: IanChilds

Differential Revision: D70855390

fbshipit-source-id: 162053e32e92e303ff80b4df61b95e7f81b3bf59